# contracts/openapi/longhub-cloud-v1.yaml
openapi: 3.1.0
info:
  title: LongHub Cloud API
  description: 龙枢云台控制面 REST 契约 V1（冻结契约草案）。破坏性变更必须创建新主版本或兼容过渡层。
  version: "1.0.0"
servers:
  - url: https://api.longhub.example.com/v1
...
security:
  - bearerAuth: []
paths:
...
  /client-releases/latest:
    get:
      tags: [release, artifact]
      summary: 获取最新的严格签名 Windows 客户端更新元数据
      description: 返回值仅供使用安装包内预置信任公钥的客户端验签；不得通过 signing-key 接口自举信任。
      operationId: getLatestClientRelease
      security: []
      parameters:
        - name: channel
          in: query
          required: false
          schema:
            type: string
            enum: [stable, beta]
            default: stable
      responses:
        "200":
          description: 最新已签名版本；当前渠道没有可用版本时 release 为 null
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [release]
                properties:
                  release:
                    oneOf:
                      - $ref: "#/components/schemas/SignedClientUpdateMetadata"
                      - type: "null"
        default:
          $ref: "#/components/responses/Error"
  /client-releases/signing-key:
    get:
      tags: [artifact]
      summary: 获取当前客户端更新签名公钥，仅供运维核对与轮换发现
      description: Desktop 不得信任本响应建立根信任；正式信任公钥必须在构建时预置。
      operationId: getClientUpdateSigningKey
      security: []
      responses:
        "200":
          description: 当前客户端更新签名公钥
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientUpdateSigningKey"
        default:
          $ref: "#/components/responses/Error"
...
  /admin/client-releases:
    get:
      tags: [release, artifact]
      summary: 管理面列出已签名客户端安装包
      operationId: listAdminClientReleases
      responses:
        "200":
          description: 已签名客户端发布列表及上传审计字段
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [releases]
                properties:
                  releases:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminClientRelease"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [release, artifact]
      summary: 管理面流式上传并签署不可覆盖的 Windows 安装包
      operationId: publishClientRelease
      parameters:
        - name: version
          in: query
          required: true
          schema: { type: string, pattern: '^(0|[1-9]\d{0,8})\.(0|[1-9]\d{0,8})\.(0|[1-9]\d{0,8})$' }
        - name: filename
          in: query
          required: true
          description: 必须精确等于 LongHub-Setup-{version}.exe
          schema: { type: string, pattern: '^LongHub-Setup-\d+\.\d+\.\d+\.exe$' }
        - name: channel
          in: query
          required: false
          schema: { type: string, enum: [stable, beta], default: stable }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              maxLength: 1073741824
      responses:
        "201":
          description: 已计算大小与 SHA-256、分配单调序列并签署元数据
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [release]
                properties:
                  release:
                    $ref: "#/components/schemas/AdminClientRelease"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ClientUpdateManifest:
      type: object
      additionalProperties: false
      description: longhub/client-update/v2 签名载荷；字段顺序无关，按共享 canonical JSON 实现签名。
      required: [schema_version, sequence, version, channel, platform, arch, filename, size, sha256, url_path, published_at, rollback_data_strategy, rollout]
      properties:
        schema_version: { type: string, const: longhub/client-update/v2 }
        sequence: { type: integer, minimum: 1, maximum: 9007199254740991 }
        version: { type: string, pattern: '^(0|[1-9]\d{0,8})\.(0|[1-9]\d{0,8})\.(0|[1-9]\d{0,8})$' }
        channel: { type: string, enum: [stable, beta] }
        platform: { type: string, const: win32 }
        arch: { type: string, const: x64 }
        filename: { type: string, pattern: '^LongHub-Setup-\d+\.\d+\.\d+\.exe$' }
        size: { type: integer, minimum: 1, maximum: 1073741824 }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        url_path:
          type: string
          pattern: '^/downloads/LongHub-Setup-\d+\.\d+\.\d+\.exe$'
        published_at: { type: string, format: date-time }
        rollback_data_strategy:
          type: string
          enum: [snapshot_required, backward_compatible]
          description: snapshot_required 表示程序降级前必须恢复安装前状态；backward_compatible 仅可在迁移兼容性已证明时使用。
        rollout:
          $ref: "#/components/schemas/ClientUpdateRollout"
    ClientUpdateRollout:
      type: object
      additionalProperties: false
      description: 签名保护的暂停与确定性灰度策略；同一版本调整比例时 seed 保持不变。
      required: [status, basis_points, seed, updated_at]
      properties:
        status: { type: string, enum: [active, paused] }
        basis_points: { type: integer, minimum: 0, maximum: 10000 }
        seed: { type: string, pattern: '^[a-f0-9]{64}$' }
        updated_at: { type: string, format: date-time }
    SignedClientUpdateMetadata:
      type: object
      additionalProperties: false
      description: Ed25519 签名覆盖用途域 longhub-client-update-v2、换行符及完整 canonical manifest；不覆盖任何响应外层字段。
      required: [manifest, signature_key_id, signature]
      properties:
        manifest:
          $ref: "#/components/schemas/ClientUpdateManifest"
        signature_key_id: { type: string, pattern: '^[a-zA-Z0-9._-]{1,128}$' }
        signature: { type: string, minLength: 1 }
    ClientUpdateSigningKey:
      type: object
      additionalProperties: false
      required: [purpose, key_id, public_key_pem]
      properties:
        purpose: { type: string, const: longhub-client-update-v2 }
        key_id: { type: string, pattern: '^[a-zA-Z0-9._-]{1,128}$' }
        public_key_pem: { type: string }
    AdminClientRelease:
      type: object
      additionalProperties: false
      required: [manifest, signature_key_id, signature, uploaded_by, uploaded_at, rollout_updated_by, rollout_updated_at, url]
      properties:
        manifest:
          $ref: "#/components/schemas/ClientUpdateManifest"
        signature_key_id: { type: string, pattern: '^[a-zA-Z0-9._-]{1,128}$' }
        signature: { type: string, minLength: 1 }
        uploaded_by: { type: string }
        uploaded_at: { type: string, format: date-time }
        rollout_updated_by: { type: string }
        rollout_updated_at: { type: string, format: date-time }
        url:
          type: string
          pattern: '^/downloads/LongHub-Setup-\d+\.\d+\.\d+\.exe$'
...
    ApiError:
      type: object
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details: {}
...